python-archive/convertions/requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
decorator==4.4.2
et-xmlfile==1.1.0
gTTS==2.5.3
html2text==2024.2.26
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
lxml==5.3.0
Markdown==3.6
markdown2==2.5.0
moviepy==1.0.3
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
pdfkit==1.0.0
pillow==10.3.0
proglog==0.1.10
pydub==0.25.1
PyMuPDF==1.24.4
PyMuPDFb==1.24.3
pypandoc==1.13
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
six==1.16.0
SpeechRecognition==3.10.4
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
